pyproject.toml
[project]
name = "ttscli"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "google-genai>=0.8",
    "elevenlabs>=1.13",
    "httpx>=0.27",
    "pydub>=0.25",
    "mutagen>=1.47",
    "audioop-lts>=0.2.2; python_version>='3.13'",
]

[project.scripts]
ttscli = "ttscli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ttscli"]
